Make stats links be the whole block rather than just the number
This has been bothering me for a while, and since the rss feed link looked weird as a blue link like the numbers, I decided to change it. The stats block looks the same now but the whole line is a link. The numbers are really hard to click on. This should be a big help for mobile as well.